requirements.txt
asgiref==3.5.2
astroid==2.4.2
certifi==2020.11.8
cffi==1.15.1
chardet==3.0.4
colorama==0.4.4
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.14
django-allauth==0.52.0
django-auditlog==2.2.2
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-countries==7.2.1
django-filter==21.1
django-import-export==2.8.0
django-js-asset==1.2.2
django-phone-field==1.8.1
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==2.10
import-export==0.2.67.dev6
isort==5.6.4
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.1.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.0.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2020.4
PyYAML==6.0
qrcode==7.3.1
requests==2.25.0
requests-oauthlib==1.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.2.1
toml==0.10.2
twilio==7.16.1
typed-ast==1.4.3
typing_extensions==4.3.0
urllib3==1.26.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
